requirements.txt
Pillow>=10.0.0
rawpy>=0.18.0
PyYAML>=6.0
textual>=0.40.0
